requirements.txt
chromadb==1.3.4
groq==0.33.0
mpmath==1.3.0
pandas==2.3.3
python-dotenv==1.2.1
semantic_router==0.1.11
streamlit==1.51.0
beautifulsoup4==4.14.2
lxml==6.0.2
pandas==2.3.3
selenium==4.38.0
